+++ lib/src/analyzers/lint_analyzer/reporters/reporters_list/console/lint_console_reporter.dart
@@ -49,7 +49,7 @@ class LintConsoleReporter
     }
 
     if (!hasReportData) {
-      output.writeln('No issues found!');
+      output.writeln('${okPen('✔')} no issues found!');
     }
   }